requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
graphene-django==3.0.0
django-graphql-auth==0.3.16
# graphql-auth==3.0.0
django-graphql-jwt==0.3.4
